refactor(web): remove dead vanilla-TS UI after Svelte port
The Svelte 5 port (a9047bb) left the original imperative-DOM modules in
place, dead but still on disk. Delete them now that the Svelte path is
verified (svelte-check 0 err, vitest 36/36, build:offline green, E2E 6/6).

Removed: index.ts, app.ts, and components/{compose-modal,explore,feed,
key-export-modal,notifications,onboarding,post-card,profile,right-panel,
settings,sidebar,thread}.ts — 2970 LOC.

Kept: repost-menu.ts (faithful extract imported by PostCard.svelte). The
shared view-model types that the deleted component factories defined
(SidebarView/ExploreItem/NotificationRecord) move to a new
components/view-types.ts; live .svelte imports repointed there. The
framework-agnostic layer (freenet-api.ts WS hub, delegate-api, identity,
shard-key, types, utils, mock-data, theme, branding) is untouched.
