--- web/src/index.ts
@@ -1,585 +1,0 @@
-import "./scss/styles.scss";
-import { APP_NAME, APP_LOGO_URL } from "./branding";
-import { initTheme } from "./theme";
-import { createApp } from "./app";
-import { FreenetConnection, type LikeState, type RepostState, type QuoteState } from "./freenet-api";
-import { Post } from "./types";
-import { createOnboarding } from "./components/onboarding";
-import { createPostCard } from "./components/post-card";
-import {
-  getIdentity,
-  createIdentity,
-  applyDelegateIdentity,
-  connectDelegate,
-  requestIdentityFromDelegate,
-  onIdentityExported,
-  Identity,
-} from "./identity";
-import { showKeyExportModal } from "./components/key-export-modal";
-import { parseDelegateResponse } from "./delegate-api";
-import { DelegateResponse } from "@freenetorg/freenet-stdlib";
-import { MOCK_POSTS } from "./mock-data";
-
-document.title = APP_NAME;
-
-// Apply persisted theme synchronously, BEFORE any render — avoids FOUC and
-// guarantees onboarding/splash also respect saved preference.
-initTheme();
-
-// Editorial design tokens are intentionally locked at boot (not user-tweakable).
-// The handoff-bundle design tooling exposed sliders for these; we ship the
-// validated combination directly. See _raven.scss for matching selectors.
-const root = document.documentElement;
-root.dataset.size = "regular";
-root.dataset.actions = "friendly";
-root.dataset.surface = "warm";
-root.dataset.avatars = "ink";
-
-// Show secret key in modal whenever delegate replies to ExportIdentity.
-onIdentityExported((secretKey) => showKeyExportModal(secretKey));
-
-const appRoot = document.getElementById("app")!;
-
-// ---------------------------------------------------------------------------
-// State
-// ---------------------------------------------------------------------------
-let localPosts: Post[] = [];
-const knownPostIds = new Set<string>();
-// Public-timeline (global-index) buffer. Independent of identity: it feeds the
-// logged-out landing feed AND the logged-in Home → Discover tab. Mirrors the
-// localPosts / knownPostIds pair above.
-let globalPosts: Post[] = [];
-const globalPostIds = new Set<string>();
-let appElement: HTMLElement | null = null;
-// The read-only public timeline mounted UNDER the onboarding overlay so a fresh
-// (logged-out) visitor lands on live network posts instead of a blank page.
-// Created lazily by showOnboarding(); torn down by renderApp() once identity is
-// known and the full app takes over.
-let landingElement: (HTMLElement & { updateGlobalPosts: (posts: Post[]) => void }) | null = null;
-let appRendered = false;
-let delegateTimeoutId: ReturnType<typeof setTimeout> | null = null;
-// onStatusChange("connected") fires on every WebSocket (re)open. The global
-// index read/subscribe must start exactly once — a reconnect re-issuing the GET
-// would re-run the snapshot merge (harmless now that it merges) and, worse,
-// re-subscribe to the singleton key. Mirror the user-shard's one-shot init
-// guard (userShardInitOwner) with a simple latch.
-let globalIndexStarted = false;
-
-// ---------------------------------------------------------------------------
-// Feed helpers
-// ---------------------------------------------------------------------------
-// createApp() attaches updatePosts on the app element itself so updates keep
-// flowing even when the visible screen is profile/explore/notifications and
-// the .feed-column DOM node is detached. The cached feed inside app.ts always
-// receives the new list; when the user navigates back to Home it re-renders
-// from the latest state, preserving the optimistic-then-reconcile contract
-// from PRs #36/#37/#38.
-type AppEl = HTMLElement & { updatePosts: (posts: Post[]) => void };
-
-function refreshFeed(): void {
-  (appElement as AppEl | null)?.updatePosts?.(localPosts);
-}
-
-// app.ts exposes updateGlobalPosts on the app element (the UI agent adds it) to
-// feed the Home → Discover tab. We forward defensively via optional-chaining so
-// this works regardless of whether that surface has landed yet. The same buffer
-// also drives the pre-auth landing feed when it is mounted.
-type GlobalAppEl = HTMLElement & { updateGlobalPosts?: (posts: Post[]) => void };
-
-function refreshGlobalFeed(): void {
-  (appElement as GlobalAppEl | null)?.updateGlobalPosts?.(globalPosts);
-  landingElement?.updateGlobalPosts(globalPosts);
-}
-
-// ---------------------------------------------------------------------------
-// Pre-auth landing feed (read-only public timeline)
-// ---------------------------------------------------------------------------
-// A minimal, read-only list of global-index posts shown behind the onboarding
-// overlay. Intentionally NOT the full createApp(): logged-out visitors have no
-// delegate/identity, so there is no compose/like/repost/quote — those actions
-// require signing. Mounting this into appRoot keeps the page from being blank
-// while the onboarding overlay (a separate document.body element) sits on top.
-function createLandingFeed(): HTMLElement & {
-  updateGlobalPosts: (posts: Post[]) => void;
-} {
-  const wrap = document.createElement("main");
-  wrap.className = "feed-column screen landing-feed";
-
-  const list = document.createElement("div");
-  // Reuse the in-app feed's styled list class so the landing list inherits the
-  // same spacing/borders (.feed-list has no SCSS rules; .feed__posts does).
-  list.className = "feed__posts";
-  wrap.appendChild(list);
-
-  const el = wrap as HTMLElement & {
-    updateGlobalPosts: (posts: Post[]) => void;
-  };
-  el.updateGlobalPosts = (posts: Post[]) => {
-    list.replaceChildren();
-    // Read-only: pass no callbacks so cards have no actionable affordances for
-    // a logged-out visitor.
-    for (const post of posts) list.appendChild(createPostCard(post));
-  };
-  return el;
-}
-
-// ---------------------------------------------------------------------------
-// Render the app (called once identity is known)
-// ---------------------------------------------------------------------------
-function renderApp(identity: Identity): void {
-  if (appRendered) return;
-  appRendered = true;
-  if (delegateTimeoutId) { clearTimeout(delegateTimeoutId); delegateTimeoutId = null; }
-
-  // Remove onboarding/splash if present
-  document.querySelector(".onboarding-overlay")?.remove();
-  document.querySelector(".splash-screen")?.remove();
-  // The full app replaces the read-only landing feed; drop it so we don't stack
-  // two feeds in appRoot.
-  landingElement?.remove();
-  landingElement = null;
-
-  connection.setUser(identity.publicKey, identity.displayName, identity.handle);
-
-  appElement = createApp({
-    publish: (content: string, shareToGlobal: boolean) => {
-      connection.publishPost(content, shareToGlobal).then((ok) => {
-        if (ok) {
-          console.log("[freenet] Post published");
-          setTimeout(() => connection.loadState(), 300);
-        }
-      });
-    },
-    like: (postId: string, liked: boolean) => {
-      connection.likePost(postId, liked).then((ok) => {
-        if (!ok) {
-          console.warn("[freenet] Like not sent (no delegate / thread shard)");
-          // Revert the optimistic toggle: re-render rebuilds the card from
-          // localPosts (the last authoritative state), discarding the toggle.
-          refreshFeed();
-        }
-      });
-    },
-    repost: (postId: string, reposted: boolean) => {
-      connection.repostPost(postId, reposted).then((ok) => {
-        if (!ok) {
-          console.warn("[freenet] Repost not sent (no delegate / thread shard)");
-          // Revert the optimistic toggle from localPosts.
-          refreshFeed();
-        }
-      });
-    },
-    quote: (postId: string, content: string) => {
-      connection.quotePost(postId, content).then((ok) => {
-        if (!ok) {
-          console.warn("[freenet] Quote not sent (no delegate / shard)");
-        } else {
-          // The quote post lands on the owner's user shard; refresh the feed so
-          // it shows up (same pattern as a normal publish).
-          setTimeout(() => connection.loadState(), 300);
-        }
-      });
-    },
-  });
-  appRoot.appendChild(appElement);
-
-  // Load posts now that app is rendered
-  connection.loadState();
-  // Hand any already-buffered public-timeline posts to the app's Discover tab
-  // (they may have arrived while the landing feed was showing). Live updates
-  // continue to flow via refreshGlobalFeed().
-  refreshGlobalFeed();
-}
-
-// ---------------------------------------------------------------------------
-// Show loading splash
-// ---------------------------------------------------------------------------
-function showSplash(): void {
-  const splash = document.createElement("div");
-  splash.className = "splash-screen";
-  splash.style.cssText =
-    "display:flex;align-items:center;justify-content:center;height:100vh;";
-  splash.innerHTML = `
-    <div style="text-align:center;display:flex;flex-direction:column;align-items:center;gap:16px">
-      <img src="${APP_LOGO_URL}" alt="${APP_NAME} logo" draggable="false" style="width:72px;height:72px;object-fit:contain;user-select:none">
-      <div style="display:flex;align-items:center;gap:8px;font-family:var(--font-mono);font-size:9.5px;letter-spacing:0.14em;text-transform:uppercase;color:var(--ink-3)">
-        <span class="live-dot"></span>
-        <span>Connecting to Freenet</span>
-      </div>
-    </div>
-  `;
-  appRoot.appendChild(splash);
-}
-
-
-// ---------------------------------------------------------------------------
-// Connection
-// ---------------------------------------------------------------------------
-const connection = new FreenetConnection({
-  onPostsLoaded: (posts: Post[]) => {
-    console.log(`[freenet] Loaded ${posts.length} posts from network`);
-    localPosts = posts;
-    knownPostIds.clear();
-    for (const post of posts) knownPostIds.add(post.id);
-    refreshFeed();
-  },
-  onNewPost: (post: Post) => {
-    if (knownPostIds.has(post.id)) return;
-    knownPostIds.add(post.id);
-    localPosts = [post, ...localPosts];
-    refreshFeed();
-  },
-  onGlobalPostsLoaded: (posts: Post[]) => {
-    console.log(`[freenet] Loaded ${posts.length} public-timeline posts`);
-    // MERGE, do not replace. The global index is multi-writer, and the
-    // subscribe fires right after the GET is queued — so a live delta
-    // (onNewGlobalPost) can land in globalPosts BEFORE this snapshot arrives.
-    // If the node served a state that predates that share, a blind
-    // `globalPosts = posts` would silently drop the just-seen post forever
-    // (its id also purged from the dedup set, so no later delta re-adds it).
-    // Keep any buffered post whose id is not in the snapshot, union, re-sort
-    // newest-first, and rebuild the dedup set from the union. (The user-shard
-    // loadState path overwrites safely only because the owner is the sole
-    // writer; the public firehose has no such guarantee.)
-    const snapshotIds = new Set(posts.map((p) => p.id));
-    const survivors = globalPosts.filter((p) => !snapshotIds.has(p.id));
-    globalPosts = [...posts, ...survivors].sort(
-      (a, b) => b.timestamp.getTime() - a.timestamp.getTime(),
-    );
-    globalPostIds.clear();
-    for (const post of globalPosts) globalPostIds.add(post.id);
-    refreshGlobalFeed();
-  },
-  onNewGlobalPost: (post: Post) => {
-    if (globalPostIds.has(post.id)) return;
-    globalPostIds.add(post.id);
-    // Insert then re-sort newest-first to preserve the buffer invariant the GET
-    // snapshot establishes. A live delta is normally the newest post (prepend
-    // alone would suffice), but the global index is multi-writer and deltas can
-    // arrive out of timestamp order — sorting keeps Discover/landing ordering
-    // correct regardless. (The single-writer user-shard onNewPost can prepend
-    // safely; the public firehose cannot assume monotonic arrival.)
-    globalPosts = [post, ...globalPosts].sort(
-      (a, b) => b.timestamp.getTime() - a.timestamp.getTime(),
-    );
-    refreshGlobalFeed();
-  },
-  onStatusChange: (status) => {
-    console.log(`[freenet] Status: ${status}`);
-
-    if (status === "connected") {
-      // Start the public-timeline read path independent of identity, so even a
-      // logged-out visitor sees live network posts on the landing feed.
-      // loadGlobalIndex() also subscribes internally (load → subscribe, mirror
-      // of initUserShard). It is fire-and-forget (it swallows its own errors)
-      // and runs before the delegate wiring so it never blocks or races the
-      // identity flow. Guarded to once per page so a reconnect does not
-      // re-subscribe to the singleton key.
-      if (!globalIndexStarted) {
-        globalIndexStarted = true;
-        connection.loadGlobalIndex();
-      }
-
-      // Wire delegate and request identity
-      wireDelegateAndRequestIdentity();
-    }
-
-    if (status === "disconnected" || status === "error") {
-      // No Freenet node — show onboarding with in-memory identity
-      if (!appRendered) {
-        showOnboarding();
-      }
-    }
-  },
-  onLikeUpdated: (like: LikeState) => {
-    // Authoritative like aggregate from the post's thread shard — reconcile the
-    // optimistic UI with real state and re-render.
-    const post = localPosts.find((p) => p.id === like.postId);
-    if (!post) return;
-    post.likes = like.count;
-    post.liked = like.likedByMe;
-    refreshFeed();
-  },
-  onRepostUpdated: (repost: RepostState) => {
-    // Authoritative repost aggregate from the post's thread shard — reconcile
-    // the optimistic UI with real state and re-render. Mirror of onLikeUpdated.
-    const post = localPosts.find((p) => p.id === repost.postId);
-    if (!post) return;
-    post.reposts = repost.count;
-    post.reposted = repost.repostedByMe;
-    refreshFeed();
-  },
-  onQuoteUpdated: (quote: QuoteState) => {
-    // Authoritative quote-repost count from the post's thread shard.
-    const post = localPosts.find((p) => p.id === quote.postId);
-    if (!post) return;
-    post.quotes = quote.count;
-    refreshFeed();
-  },
-  onDelegateResponse: (response: DelegateResponse) => {
-    const payloads = parseDelegateResponse(response);
-    for (const payload of payloads) {
-      console.log("[delegate] Response:", payload);
-
-      if (applyDelegateIdentity(payload)) {
-        const identity = getIdentity()!;
-        connection.setUser(identity.publicKey, identity.displayName, identity.handle);
-
-        if (!appRendered) {
-          renderApp(identity);
-        }
-        return;
-      }
-
-      // A signed post came back from the delegate — finish publishing it.
-      const signed = payload as {
-        type?: string;
-        nonce?: string;
-        post_id?: string;
-        signature?: string;
-        public_key?: string;
-      };
-      if (
-        signed.type === "Signed" &&
-        signed.nonce &&
-        signed.post_id &&
-        signed.signature &&
-        signed.public_key
-      ) {
-        connection
-          .completePublish({
-            nonce: signed.nonce,
-            post_id: signed.post_id,
-            signature: signed.signature,
-            public_key: signed.public_key,
-          })
-          .catch((e) => console.error("[delegate] completePublish failed:", e));
-        return;
-      }
-
-      // A signed like came back — fold it into the post's thread shard.
-      const signedLike = payload as {
-        type?: string;
-        nonce?: string;
-        root_post_id?: string;
-        signer_pubkey?: string;
-        seq?: number;
-        liked?: boolean;
-        signature?: string;
-      };
-      if (
-        signedLike.type === "SignedLike" &&
-        signedLike.nonce &&
-        signedLike.root_post_id &&
-        signedLike.signer_pubkey &&
-        typeof signedLike.seq === "number" &&
-        typeof signedLike.liked === "boolean" &&
-        signedLike.signature
-      ) {
-        connection
-          .completeLike({
-            nonce: signedLike.nonce,
-            root_post_id: signedLike.root_post_id,
-            signer_pubkey: signedLike.signer_pubkey,
-            seq: signedLike.seq,
-            liked: signedLike.liked,
-            signature: signedLike.signature,
-          })
-          .then((ok) => {
-            // completeLike already fires a best-effort re-GET to reconcile, but
-            // that depends on a second round-trip succeeding (and the network
-            // condition that failed the UPDATE likely fails the GET too). On a
-            // hard false, revert the optimistic toggle from localPosts directly
-            // so the like never sticks on the card. Mirrors the likePost(false)
-            // and dropPendingLike paths.
-            if (!ok) refreshFeed();
-          })
-          .catch((e) => console.error("[delegate] completeLike failed:", e));
-        return;
-      }
-
-      // A signed repost came back — fold it into the post's thread shard.
-      const signedRepost = payload as {
-        type?: string;
-        nonce?: string;
-        root_post_id?: string;
-        signer_pubkey?: string;
-        seq?: number;
-        reposted?: boolean;
-        signature?: string;
-      };
-      if (
-        signedRepost.type === "SignedRepost" &&
-        signedRepost.nonce &&
-        signedRepost.root_post_id &&
-        signedRepost.signer_pubkey &&
-        typeof signedRepost.seq === "number" &&
-        typeof signedRepost.reposted === "boolean" &&
-        signedRepost.signature
-      ) {
-        connection
-          .completeRepost({
-            nonce: signedRepost.nonce,
-            root_post_id: signedRepost.root_post_id,
-            signer_pubkey: signedRepost.signer_pubkey,
-            seq: signedRepost.seq,
-            reposted: signedRepost.reposted,
-            signature: signedRepost.signature,
-          })
-          .then((ok) => {
-            // On a hard false, revert the optimistic toggle from localPosts so
-            // the repost never sticks on the card (mirrors the like path).
-            if (!ok) refreshFeed();
-          })
-          .catch((e) => console.error("[delegate] completeRepost failed:", e));
-        return;
-      }
-
-      // A signed quote ref came back — fold it into the quoted post's thread shard.
-      const signedQuote = payload as {
-        type?: string;
-        nonce?: string;
-        root_post_id?: string;
-        signer_pubkey?: string;
-        quote_post_id?: string;
-        signature?: string;
-      };
-      if (
-        signedQuote.type === "SignedQuoteRef" &&
-        signedQuote.nonce &&
-        signedQuote.root_post_id &&
-        signedQuote.signer_pubkey &&
-        signedQuote.quote_post_id &&
-        signedQuote.signature
-      ) {
-        connection
-          .completeQuoteRef({
-            nonce: signedQuote.nonce,
-            root_post_id: signedQuote.root_post_id,
-            signer_pubkey: signedQuote.signer_pubkey,
-            quote_post_id: signedQuote.quote_post_id,
-            signature: signedQuote.signature,
-          })
-          .catch((e) => console.error("[delegate] completeQuoteRef failed:", e));
-        return;
-      }
-
-      // Check for an error from the delegate.
-      const p = payload as { type?: string; message?: string; nonce?: string };
-      if (p.type === "Error") {
-        // If the error carries a nonce it came from a failed SignPost,
-        // SignLike, or SignRepost — drop exactly that stranded pending action.
-        // Errors without a nonce (GetIdentity, Export, …) leave the queues
-        // untouched.
-        if (p.nonce) {
-          connection.dropPendingPost(p.nonce);
-          // A dropped pending like/repost leaves an un-acked optimistic toggle
-          // on its card — re-render from localPosts to revert it.
-          if (connection.dropPendingLike(p.nonce)) refreshFeed();
-          if (connection.dropPendingRepost(p.nonce)) refreshFeed();
-          connection.dropPendingQuoteRef(p.nonce);
-        }
-        if (p.message?.includes("no identity")) {
-          console.log("[identity] No identity in delegate — show onboarding");
-          if (!appRendered) {
-            showOnboarding();
-          }
-        } else {
-          console.warn("[delegate] Error:", p.message);
-        }
-        return;
-      }
-    }
-  },
-});
-
-// ---------------------------------------------------------------------------
-// Delegate wiring
-// ---------------------------------------------------------------------------
-async function wireDelegateAndRequestIdentity(): Promise<void> {
-  const api = connection.wsApi;
-  if (!api || !__DELEGATE_KEY__) {
-    // No delegate configured — fallback to onboarding
-    if (!appRendered) showOnboarding();
-    return;
-  }
-
-  try {
-    // Use pre-decoded bytes injected at build time (avoids base58 decode in sandbox)
-    const keyBytes = __DELEGATE_KEY_BYTES__;
-    const codeHashBytes = __DELEGATE_CODE_HASH_BYTES__;
-    if (!keyBytes || keyBytes.length !== 32) {
-      console.warn("[identity] Invalid delegate key bytes, length:", keyBytes?.length);
-      if (!appRendered) showOnboarding();
-      return;
-    }
-    if (!codeHashBytes || codeHashBytes.length !== 32) {
-      console.warn("[identity] Invalid delegate code_hash bytes, length:", codeHashBytes?.length);
-      if (!appRendered) showOnboarding();
-      return;
-    }
-
-    connectDelegate(api, keyBytes, codeHashBytes);
-    console.log(`[identity] Delegate wired (key: ${keyBytes.length}b, code_hash: ${codeHashBytes.length}b)`);
-
-    requestIdentityFromDelegate();
-
-    delegateTimeoutId = setTimeout(() => {
-      if (!appRendered) {
-        console.log("[identity] Delegate timeout — showing onboarding");
-        showOnboarding();
-      }
-      delegateTimeoutId = null;
-    }, 5000);
-  } catch (e) {
-    console.warn("[identity] Failed to wire delegate:", e);
-    if (!appRendered) showOnboarding();
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Onboarding
-// ---------------------------------------------------------------------------
-function showOnboarding(): void {
-  if (appRendered) return;
-  // Remove splash
-  document.querySelector(".splash-screen")?.remove();
-
-  // Mount the read-only public timeline UNDER the overlay so the page isn't
-  // blank while logged out. The overlay (appended to document.body below) is a
-  // full-screen element that layers on top of appRoot. Seed it with whatever
-  // global posts have already arrived; refreshGlobalFeed() keeps it live.
-  if (!landingElement) {
-    landingElement = createLandingFeed();
-    appRoot.appendChild(landingElement);
-    landingElement.updateGlobalPosts(globalPosts);
-  }
-
-  const onboarding = createOnboarding((displayName: string, secretKey?: string) => {
-    const identity = createIdentity(displayName, secretKey);
-    renderApp(identity);
-  });
-  document.body.appendChild(onboarding);
-}
-
-// ---------------------------------------------------------------------------
-// Start
-// ---------------------------------------------------------------------------
-if (__OFFLINE_MODE__) {
-  // Offline mode: skip WebSocket / delegate, render immediately with mock data.
-  // Used for CI / Playwright / no-node previews.
-  console.log("[offline] Booting in offline mode with mock data");
-  const identity = createIdentity("Offline Demo");
-  renderApp(identity);
-  // renderApp triggers connection.loadState() which is a no-op without a connect().
-  // Push mock posts into the feed once the app is mounted.
-  setTimeout(() => {
-    const feed = appElement?.querySelector(".feed-column") as
-      | (HTMLElement & { updatePosts: (posts: Post[]) => void })
-      | null;
-    feed?.updatePosts(MOCK_POSTS);
-  }, 0);
-} else {
-  showSplash();
-  connection.connect();
-}
